types: eliminate no-any-return baseline
Normalize declared scalar return contracts and cast dynamic boundaries without changing runtime behavior.
Refs: #4

+++ src/boofun/analysis/communication_complexity.py
@@ -260,7 +260,7 @@ def discrepancy(f: "BooleanFunction") -> float:
     N = rows * cols
     spectral_norm: float = float(np.max(svdvals(M_pm.astype(float))))
 
-    return spectral_norm / N
+    return float(spectral_norm / N)
 
 
 def deterministic_cc(f: "BooleanFunction") -> dict[str, Any]:
